Remove window-level scopes on window close
